Fix windowing
