+++ sunscreen_math/src/opencl_impl/multiexp.rs
@@ -100,3 +100,154 @@ pub fn multiscalar_multiplication(points: &GpuRistrettoPointVec, scalars: &GpuSc
 
      todo!();
 }
+
+#[cfg(test)]
+mod tests {
+    use rand::thread_rng;
+
+    use crate::ScalarVec;
+
+    use super::*;
+
+    #[test]
+    fn can_get_msm_scalar_windows() {
+        let a = (0..4567)
+            .map(|_| Scalar::random(&mut thread_rng()))
+            .collect::<Vec<_>>();
+
+        let a_gpu = ScalarVec::new(&a);
+        let runtime = Runtime::get();
+
+        for window_bits in 1..33 {
+            const SCALAR_BITS: usize = 8 * std::mem::size_of::<Scalar>();
+
+            let num_windows = if SCALAR_BITS % window_bits == 0 {
+                SCALAR_BITS / window_bits
+            } else {
+                SCALAR_BITS / window_bits + 1
+            };
+
+            let mut windows_gpu = runtime.alloc::<u32>(num_windows * a.len());
+
+            runtime.run_kernel(
+                "test_get_scalar_windows",
+                &[
+                    (&a_gpu.data).into(),
+                    (&windows_gpu).into(),
+                    (window_bits as u32).into(),
+                    (a.len() as u32).into(),
+                ],
+                &Grid::from([(a.len(), 256), (num_windows, 1), (1, 1)]),
+            );
+
+            // The windows buffer's contents have changed since running the kernel, so we need
+            // to remap the host address on some (e.g. Nvidia) platforms.
+            windows_gpu.remap();
+
+            let windows = windows_gpu.iter().cloned().collect::<Vec<_>>();
+
+            for i in 0..a.len() {
+                let mut actual = Scalar::zero();
+                let mut radix = Scalar::one();
+
+                for j in 0..num_windows {
+                    let cur_window = windows[j * a.len() + i];
+                    assert!((cur_window as u64) < (0x1u64 << window_bits as u64));
+
+                    actual += Scalar::from(cur_window) * radix;
+                    radix *= Scalar::from(0x1u64 << window_bits as u64);
+                }
+
+                assert_eq!(actual, a[i]);
+            }
+        }
+    }
+
+    #[test]
+    fn can_fill_coo_matrix() {
+        let a = (0..4567)
+            .map(|_| Scalar::random(&mut thread_rng()))
+            .collect::<Vec<_>>();
+
+        let a_gpu = ScalarVec::new(&a);
+        let runtime = Runtime::get();
+
+        let window_bits = 15;
+
+        const SCALAR_BITS: usize = 8 * std::mem::size_of::<Scalar>();
+
+        let num_windows = if SCALAR_BITS % window_bits == 0 {
+            SCALAR_BITS / window_bits
+        } else {
+            SCALAR_BITS / window_bits + 1
+        };
+
+        let mut windows_gpu = runtime.alloc::<u32>(num_windows * a.len());
+
+        runtime.run_kernel(
+            "test_get_scalar_windows",
+            &[
+                (&a_gpu.data).into(),
+                (&windows_gpu).into(),
+                (window_bits as u32).into(),
+                (a.len() as u32).into(),
+            ],
+            &Grid::from([(a.len(), 256), (num_windows, 1), (1, 1)]),
+        );
+
+        // The windows buffer's contents have changed since running the kernel, so we need
+        // to remap the host address on some (e.g. Nvidia) platforms.
+        windows_gpu.remap();
+
+        let windows = windows_gpu.iter().cloned().collect::<Vec<_>>();
+
+        const NUM_THREADS: usize = 4;
+
+        let mut coo_data = runtime.alloc(a.len() * num_windows);
+        let mut coo_row_idx = runtime.alloc(a.len() * num_windows);
+        let mut coo_col_idx = runtime.alloc(a.len() * num_windows);
+
+        runtime.run_kernel(
+            "fill_coo_matrix",
+            &[
+                (&a_gpu.data).into(),
+                (&coo_data).into(),
+                (&coo_row_idx).into(),
+                (&coo_col_idx).into(),
+                (window_bits as u32).into(),
+                (a.len() as u32).into(),
+            ],
+            &Grid::from([(NUM_THREADS, 2), (num_windows, 1), (1, 1)]),
+        );
+
+        coo_data.remap();
+        coo_row_idx.remap();
+        coo_col_idx.remap();
+
+        let coo_data = coo_data.iter().cloned().collect::<Vec<_>>();
+        let coo_row_idx = coo_row_idx.iter().cloned().collect::<Vec<_>>();
+        let coo_col_idx = coo_col_idx.iter().cloned().collect::<Vec<_>>();
+
+        for w in 0..num_windows {
+            for r in 0..a.len() {
+                assert_eq!(coo_row_idx[a.len() * w + r], (r % NUM_THREADS) as u32);
+            }
+        }
+
+        for w in 0..num_windows {
+            let start = a.len() * w;
+            let end = start + a.len();
+
+            let window = &windows[start..end];
+            let cols = &coo_col_idx[start..end];
+
+            for r in 0..a.len() {
+                assert_eq!(window[r], cols[r]);
+            }
+        }
+
+        for i in 0..a.len() {
+            assert!(coo_data.contains(&(i as u32)));
+        }
+    }
+}
